pyproject.toml
[project]
name = "vscope"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "numpy>=2",
    "platformdirs>=4.3.8",
    "pyinstaller>=6.14.1",
    "pyqt6>=6.9.1",
    "pyqtgraph==0.13.7",
    "pyserial>=3.5",
    "pytest>=8.4.1",
    "tomli-w>=1.0.0"
]
